tiefightR - rhesus data

Six male rhesus macaques (Macaca mulatta) at an age range of 6-19 years (mean = 13.09) living in same-sexual groups of 2 - 4 at the Cognitive Neuroscience Laboratory, German Primate Centre, took part in this study. Monkeys were housed in indoor rooms equipped with a multitude of toys and wooden structures as well as natural and artificial light. The space per monkey exceeded all applicable German and European regulations (Berger et al. 2018). Indoor rooms were temperature-controlled and connected by a tunnel with rooms at ambient outdoor temperature and lighting, but protected from precipitation. On test days, monkeys had free access to water for at least 4 hours (typically much more: for definitions of access to water see (Pfefferle et al. 2018)) and received monkey chow ad libitum. On non-test days, the monkeys had free access to water and received monkey chow ad libitum, supplemented with dried fruits, fresh fruits and vegetables. The health of the monkeys was monitored daily by the animal care staff, veterinarians, and the laboratory researchers who were all highly experienced with these animals.

data(rhesus)

Format

A 240 x 7 data frame with continouse response data (amountDrank).

Examples

data(rhesus)
head(rhesus)

#>       date animalID  side fluidType amountDrank totalDrank combinationWith
#> 1 21.11.18     Iggy  left    banana         365        385            NaCl
#> 2 21.11.18     Iggy right      NaCl          20        385          banana
#> 3 21.11.18    Joshi  left     water          20        310           grape
#> 4 21.11.18    Joshi right     grape         290        310           water
#> 5 22.11.18     Iggy  left     water          70        470          banana
#> 6 22.11.18     Iggy right    banana         400        470           water